bench: guard the reference harness and repair the Java benchmark (#398)
Two defects in the benchmark harnesses, found while re-measuring every published
number against v1.0.0. **Neither touches library code** — they change what the
benchmarks *report*, not what Wickra does.

## The reference row measured a loop the optimiser could delete

`examples/rust/src/bin/throughput.rs` is the FFI-free Rust baseline that all nine
bindings in BENCHMARKS.md §3 are compared against. It called

```rust
for &price in &close {
    ind.update(price);      // result dropped
}
```

The three criterion benches in this repository all guard their loops with
`black_box`; this one — an example binary rather than a bench target — never did,
not since the benchmarks landed in #246.

Measured cost of the omission on `SMA(20)`, 200 000 bars:

| | unguarded | guarded |
|---|---|---|
| streaming | 1836 Mupd/s | **1341** Mupd/s |
| batch | 514 | 505 |

A 27% inflation, and it lands on the one row every other row is measured against.
Batch was never affected — the result array is returned, so there was nothing to
elide.

## The Java benchmark did not compile

`bindings/java/benchmarks` pinned `wickra` **0.9.6**, three releases behind, so it
had been building against an API where `Atr.batch` took a `double[]` of
timestamps. The current API takes `long[]`.

- against a current jar: `NoSuchMethodError: double[] org.wickra.Atr.batch(double[], double[], double[], double[], double[], double[])`
- against its own source: `Throughput.java:[97] incompatible types: double[] cannot be converted to long[]`

Nothing reported it because CI does not build that module. The array is now
`long[]` and the now-redundant `(long)` cast at the update call site is gone. With
that, the Java row measures again: 60 Mupd/s streaming, 190 batch.

## Deliberately not in this PR

The published §3 reference value — **380 Mupd/s streaming** — does not reproduce
either way; guarded it measures 1341. That is a question for the tables, not for
the harness, and it is one of several numbers that need re-measuring on a machine
without foreign CPU load before anything is rewritten.

+++ bindings/java/benchmarks/src/main/java/org/wickra/benchmarks/Throughput.java
@@ -55,7 +55,7 @@ public final class Throughput {
         double[] low = new double[bars];
         double[] close = new double[bars];
         double[] volume = new double[bars];
-        double[] timestamp = new double[bars];
+        long[] timestamp = new long[bars];
         for (int i = 0; i < bars; i++) {
             double mid = 100 + Math.sin(i * 0.001) * 20 + i * 1e-4;
             double c = mid + Math.sin(i * 0.05) * 2;
@@ -88,7 +88,7 @@ public final class Throughput {
                 () -> {
                     try (Atr ind = new Atr(14)) {
                         for (int i = 0; i < n; i++) {
-                            ind.update(open[i], high[i], low[i], close[i], volume[i], (long) timestamp[i]);
+                            ind.update(open[i], high[i], low[i], close[i], volume[i], timestamp[i]);
                         }
                     }
                 },
